methods/anycam_2/requirements.txt
torch
torchvision
torchaudio
transformers
opencv-python
ipython
matplotlib
rerun-sdk
pytorch-ignite
moviepy
dotdict
einops
scipy
wandb
tensorboard
xformers<=0.0.28.post3
triton<=3.1
hydra-core
pycolmap
evo
tabulate
scikit-image
pyceres
kornia
timm

projectaria-tools
